FixedSmallBugs

+++ Src/Services/Eshop.Orders/Models/OrderItem.cs
@@ -5,8 +5,8 @@
         public int Id { get; set; }
         public int ProductId { get; set; }
         public int Quantity { get; set; }
-        public int UnitPrice { get; set; }
-        public int FullPrice { get; set; }
+        public double UnitPrice { get; set; }
+        public double FullPrice { get; set; }
         public int InventoryId { get; set; }
         public int OrderId { get; set; }
         public Order Order { get; set; }
